Fix ILibGetTimeStamp() overflow on 32-bit targets
ILibGetTimeStamp() returned the wall clock in milliseconds as a long.
On 32-bit Linux a long is 32 bits, and tv_sec * 1000 overflows, which
is undefined behaviour. A debug build panics on it at startup, in
the chain watchdog thread. Release builds silently wrapped.

Fix: The function now returns long long, and its callers store the value as
long long.

+++ meshcore/agentcore.h
@@ -260,7 +260,7 @@ typedef struct MeshAgentHostContainer
 	char *meshServiceName;
 	char *displayName;
 	int serviceReserved;
-	long consoleText_timeStamp;
+	long long consoleText_timeStamp;
 	int consoleText_counter;
 	int consoleText_maxRate;
 	int autoproxy_status;
